Fix matching against any()
In #54, an extra clause was added to this predicate:
https://github.com/steveklabnik/semver/pull/54/files#diff-3b37247d04b36bc46203d275684ce18bL235

If self.predicates is empty, .all() will return true, but .any() will
return false. This means that previously correct behavior, returning
true, would now be incorrect, as any() will always be false, and
(anything && false) is false.

By explicitly checking for the empty predicate case and returning early,
all is okay again.