replace nom with semver-parser
I really like nom, but at the same time, it's very complicated for a
grammar that's not particularly complicated. I have been finding it a
bit hard to maintain.

On the other hand, I really like regular expressions. And for a grammar
as small and regular as semver, it's a nice fit. As such, I decided to
implement parsing via regexes as a semver-parser crate. It was very
quick and straightforward to do. This commit therefore removes nom, and
adds in semver-parser.

My intention is to also port version requirement strings to
semver-parser as well, but that will happen in a future commit.