Parse deprecated versions.
Due to issues like https://github.com/rust-lang/cargo/issues/3124, we've
found that the old parser was lenient in ways that the new parser isn't.
We've analyzed all the crates on crates.io, and found the patterns that
weren't working.

This commit adds support back for those versions, but marks them as
deprecated. Each of them had a different way in which they were
incorrect, but resolved to some kind of correct constraint. So the
deprecation strategy is this:

1. If a version fails to parse, check if it matches the deprecated
pattern.
2. If it does, convert it to the real pattern.
3. Return an error that it's deprecated, and the error will contain the
fixed-up version constraint.