Bugfix for rust-lang/cargo#3202
We improperly assumed that if we had a wildcard patch version, we must
have a non-wildcard minor version. This is actually incorrect.

This bug has been present in semver since at least the 0.2 series, but
`cargo build` never happend to invoke Display for Predicate. The bug
manifested in newer Cargos because in 0.5, we do call Display for
Predicate.