Add Diesel support
This code targets Diesel 1.1. I can support Diesel 1.0 if it's required,
but it will require a lot more code. Impls like this are the main focus
of 1.1.

While it'd be great if we could just have a diesel-semver crate, Rust
today requires that this be either in Diesel or Semver.

This makes the same assumptions as the serde implementations. We assume
that the values are being stored as a string, and getting them in/out is
just `FromStr` and `Display`.

I've opted to only support `Version` and `VersionReq` in this case.
Though the serde implementation supports `Identifier`, there's not an
obvious database agnostic way to store that type, and I don't think it's
nearly as commonly used as the other two.

Given that this crate supports Serde, I think it makes just as much
sense to add minimal Diesel support with the same assumptions.
Especially given that it's only ~40 lines of code. This change will
enable the removal of around a hundred lines from crates.io, and enable
a bunch of refactorings that I've been wanting to do there.