Remove most unsafe code from `identifier.rs`.
The old implementation, while being really efficient and fast, was very unsafe and had problems with miri's raw pointer tagging.

This rewrite does come at a performance cost (0.25ns -> 0.34ns in the prerelease parsing test), but I don't think that this will be noticeable anywhere.

The only remaining unsafe code is a `str::from_utf8_unchecked` to convert the inline array to a `str` efficiently.