Add BareVersion
Rust 1.56.0 has introduced the rust-version field in Cargo.toml, which[1]:

> must be a bare version number with two or three components;
> it cannot include semver operators or pre-release identifiers.

semver::Version requires three components so it's Deserialize
implementation cannot be used for bare versions.

While bare versions can be deserialized as semver::VersionReq by virtue
of being a syntactic subset, it doesn't make much sense to deserialize
strings that explicitly "cannot include semver operators" as a vector of
comparison operators.

This commit therefore introduces a new BareVersion type to accurately
represent the restrictions of the new rust-version Cargo.toml field.

Prior art includes the cargo-msrv crate which already defines its own
BareVersion type.[2] The type introduced by this commit differs in that
it's a struct instead of an enum since that allows easier access to the
major and minor components.

[1]: https://doc.rust-lang.org/cargo/reference/manifest.html#the-rust-version-field
[2]: https://github.com/foresterre/cargo-msrv/blob/v0.15.1/src/manifest/bare_version.rs#L7-L11